11 table column comparision (#14)
* #11 table column and more

* #11 tests

* #11 fix enumsortorder type

* #11 use_drop flag

* #11 use_drop

* #11 use_drop

* #11 format

* #11

* #11 MIT

* #11

* #11

* #11

* #11 fix tests and triggers

* #11 views

* #11

* #11 table_column link to view

* #11 table_column functions

* #11 gitignore

* #11

* #11

* #11 fixed some output script issues.

* #11 tests

* #11

* #11 optimisation

* #11

* #11 fix add column

* #11

* #11 final columns

+++ app/src/dump/pg_enum.rs
@@ -6,6 +6,6 @@ use sqlx::postgres::types::Oid;
 pub struct PgEnum {
     pub oid: Oid,           // Oid of Enum type
     pub enumtypid: Oid,     // Oid of the Enum type
-    pub enumsortorder: i32, // Sort order of the enum value
+    pub enumsortorder: f32, // Sort order of the enum value
     pub enumlabel: String,  // Label of the enum value
 }
